Added a --mazey option to the ep.py script
Since we all like being environmentally friendly, I think it's important
that we have an option that only prints a random maze and no other data.

For one that like doing mazes, it's a lot of paper to waste to print all
the data every time one makes a cup of tea.